Large refactoring. All libraries are moved into Framework and Utilities directories
