+++ Framework/Analysis/src/AnalysisManagerBase.cxx
@@ -1,3 +1,3 @@
-#include "AnalysisManagerBase.h"
+#include "AnalysisRUDA/AnalysisManagerBase.h"
 
 /*******************************************************************************************************************/
