+++ Utilities/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_subdirectory(Common)
+add_subdirectory(RunManager)
