+++ Utilities/Common/include/CommonRUDA/AnalysisUtils.h
