+++ Utilities/Common/src/HelperCommon.cxx
@@ -1,0 +1,1 @@
+#include "CommonRUDA/HelperCommon.h"
