+++ Utilities/Common/src/HistUtils.cxx
@@ -1,4 +1,4 @@
-#include "HistUtils.h"
+#include "CommonRUDA/HistUtils.h"
 using namespace utilities;
