+++ Utilities/RunManager/include/RunManagerRUDA/RunManager.h
@@ -3,6 +3,8 @@
 #include <iostream>
 #include <vector>
 #include <string>
+#include <map>
+#include <set>
 
 #include "TTree.h"
 #include "TFile.h"
