--- src/common/HelperCommon.cxx
@@ -1,1 +1,0 @@
-#include "HelperCommon.h"
