--- src/common/HelperHists.cxx
@@ -1,1 +1,0 @@
-#include "HelperHists.h"
