+++ tmp/CMakeLists.txt
@@ -1,0 +1,4 @@
+if(AliRoot_FOUND)
+  add_subdirectory(grid)
+  add_subdirectory(trigger)
+endif()
