+++ tmp/trigger/TriggerManager.h
